add server cache

+++ src/services/postgres/PlaylistsService.js
@@ -5,8 +5,9 @@ const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class PlaylistService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this._collaborationsService = collaborationsService;
+    this._cacheService = cacheService;
     this._pool = new Pool();
   }
 
@@ -22,23 +23,29 @@ class PlaylistService {
     if (!result.rows[0].id) {
       throw new InvariantError('tidak dapat menambahkan playliust');
     }
-
+    this._cacheService.delete(`playlist:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    const query = {
-      text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON playlists.owner=users.id LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id WHERE playlists.owner = $1 OR collaborations.user_id = $1',
-      values: [owner],
-    };
+    try {
+      const result = await this._cacheService.get(`playlist:${owner}`);
+      return JSON.parse(result);
+    } catch (e) {
+      const query = {
+        text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON playlists.owner=users.id LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id WHERE playlists.owner = $1 OR collaborations.user_id = $1',
+        values: [owner],
+      };
 
-    const result = await this._pool.query(query);
-    return result.rows;
+      const result = await this._pool.query(query);
+      await this._cacheService.set(`playlist:${owner}`, JSON.stringify(result.rows));
+      return result.rows;
+    }
   }
 
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING owner',
       values: [id],
     };
 
@@ -47,6 +54,9 @@ class PlaylistService {
     if (!result.rowCount) {
       throw new NotFoundError('Playlist tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`playlist:${owner}`);
   }
 
   async addSongToPlaylist(playlistId, songId) {
@@ -61,18 +71,24 @@ class PlaylistService {
     if (!result.rows[0].id) {
       throw new InvariantError('tidak dapat menambahkan lagu ke playlist');
     }
-
+    await this._cacheService.delete(`songs:${playlistId}`);
     return result;
   }
 
   async getSongsPlaylist(playlistId) {
-    const query = {
-      text: 'SELECT songs.id, songs.title, songs.performer FROM playlists INNER JOIN playlistsongs ON playlists.id = playlistsongs.playlist_id INNER JOIN songs ON songs.id = playlistsongs.song_id WHERE playlists.id = $1',
-      values: [playlistId],
-    };
+    try {
+      const result = await this._cacheService.get(`songs:${playlistId}`);
+      return JSON.parse(result);
+    } catch (e) {
+      const query = {
+        text: 'SELECT songs.id, songs.title, songs.performer FROM playlists INNER JOIN playlistsongs ON playlists.id = playlistsongs.playlist_id INNER JOIN songs ON songs.id = playlistsongs.song_id WHERE playlists.id = $1',
+        values: [playlistId],
+      };
 
-    const result = await this._pool.query(query);
-    return result.rows;
+      const result = await this._pool.query(query);
+      await this._cacheService.set(`songs:${playlistId}`, JSON.stringify(result.rows));
+      return result.rows;
+    }
   }
 
   async deleteSongsOnPlaylist(songId, playlistId) {
@@ -85,6 +101,8 @@ class PlaylistService {
     if (!result.rowCount) {
       throw new InvariantError('Tidak dapat menghapus lagu dari playlist');
     }
+
+    await this._cacheService.delete(`songs:${playlistId}`);
   }
 
   async verifyPlaylistOwner(playlistId, owner) {
